+++ src/commands/create/mod.rs
@@ -1,0 +1,9 @@
+mod trigger;
+mod function;
+mod view;
+mod migration;
+
+pub use migration::*;
+pub use trigger::*;
+pub use function::*;
+pub use view::*;
