+++ src/commands/create/templates/function.sql
