+++ src/commands/create/templates/trigger_function.sql
