+++ src/commands/create/templates/view.sql
